integrating interfaces

+++ Interfaces/ITransportDepotOptions.cs
@@ -4,7 +4,7 @@ public interface ITransportDepotOptions
 
 {
     public int VehicleCapacity { get; set; }
-    public int MaintenanceDuration { get; set; }
-    public int ProductionDuration { get; set; }
+    public float MaintenanceDuration { get; set; }
+    public float ProductionDuration { get; set; }
 
 }
