+++ Interfaces/ITransportStopOptions.cs
@@ -3,11 +3,11 @@
 public interface ITransportStopOptions
 
 {
-    public int EnterDistance { get; set; }
-    public int ExitDistance { get; set; }
-    public int AccessDistance { get; set; }
-    public int BoardingTime { get; set; }
-    public int ComfortFactor { get; set; }
-    public int LoadingFactor { get; set; }
+    public float EnterDistance { get; set; }
+    public float ExitDistance { get; set; }
+    public float AccessDistance { get; set; }
+    public float BoardingTime { get; set; }
+    public float ComfortFactor { get; set; }
+    public float LoadingFactor { get; set; }
 
 }
